Updates component tests
* move the configuration to own file
*update the description to make it clearer
