+++ ui/components/AccountsNotificationPanel/__tests__/SigningButton.test.tsx
@@ -3,7 +3,7 @@ import userEvent from "@testing-library/user-event"
 import { createMemoryHistory } from "history"
 import React from "react"
 import { Router } from "react-router-dom"
-import { renderWithProviders } from "../../../utils/test-utils"
+import { renderWithProviders } from "../../../tests/test-utils"
 import SigningButton from "../SigningButton"
 
 const onChange = jest.fn()
@@ -18,7 +18,7 @@ const getPreloadedState = (status: "locked" | "unlocked") => ({
 })
 
 describe("SigningButton", () => {
-  test("should go to the unlocking page after clicking when the wallet is locked", async () => {
+  test("should go to the unlocking page after being clicked when the wallet is locked", async () => {
     const history = createMemoryHistory()
     const ui = renderWithProviders(
       <Router history={history}>
@@ -34,7 +34,7 @@ describe("SigningButton", () => {
     expect(history.location.pathname).toBe("/keyring/unlock")
   })
 
-  test("should go to the root page after clicking when the wallet is unlocked", async () => {
+  test("should go to the root page after being clicked when the wallet is unlocked", async () => {
     const history = createMemoryHistory()
     const ui = renderWithProviders(
       <Router history={history}>
